UI for Student Exam Pages
I made the UI for the exam pages however it lacks lots of functionality. Right now to see it if you click on grades u can see the main exam menu page and then settings takes you to the quiz page. Actual directions from the correct buttons still needs to be added however for now ive done the design of the pages. Will focus on functionality soon